HDDS-12918. Move ranged S3 fallback to test stub
